Interpret functions - @deck.gl/json
